Fix preExecutionRules with @apollo/server V4
If using @apollo/server v4 the context value in rule execution is undefined.

This is because: "The context field has been renamed contextValue for consistency with the graphql-js API and to help differentiate from the context option of integration functions (the function which returns a context value)." as seen in https://www.apollographql.com/docs/apollo-server/migration/#graphqlresponse 

To fix this, I have added "|| requestContext.contextValue" into the preExecutionRules execution code. Shouldn't affect anything else, but it makes this aspect of the plugin compatible.

Note: I am not going to touch the postExecRules since that dives into the @graphql-authz/core package... and I don't use that feature anyway.

Would be super grateful if this could be merged ASAP so I can continue using @graphql-authz/apollo-server-plugin... I imagine the apollo update might have other issues, but this is a quick fix to get the main functionality working again.